Fixed an example to be usable for demonstration
The previous version of the code had a problem with initially deploying the contracts. Consumer could not be deployed with any wei in it and therefore could not send any value to InfoFeed in callFeed() causing a revert on try.
By adding constructor() payable {} I allowed wei to be sent to the contract on deploy. 
I've also added a balance() function that returns the balance of the contract to InfoFeed to make it possible to demonstrate the effect of callFeed()